relocate global module index back in modules.common, and also relocate all of the module index logic there
